fix(geometry): move pole-keying helpers to _pole_keying.py — un-clobber the public rasterize function (#274, CI red)
rfx/geometry/__init__.py re-exports the public FUNCTION rasterize from
csg.py while rfx/geometry/rasterize.py is also a SUBMODULE. The #274
module-level import in rfx/api/_compile.py initialized that submodule
during 'import rfx'; submodule init setattr's the module object onto
the package, clobbering the function attribute — the rcs_scattering
tutorial's 'from rfx.geometry import rasterize' then received the
module (TypeError: 'module' object is not callable; fast-suite shard 4
red). Pre-existing landmine (the NU/subgridded runners' function-scoped
submodule imports trip it too, post-run); the #274 import was merely
the first to trip it at import-rfx time. The name collision itself
deserves its own tracked issue — renaming either side is a
public-surface change out of this PR's scope.

Fix: pure code motion. _pole_key / _accumulate_pole_mask /
_spec_from_pole_masks move VERBATIM to the collision-free
rfx/geometry/_pole_keying.py; rasterize.py imports them from there
(internal call sites unchanged); _compile.py imports from
rfx.geometry._pole_keying — restoring exactly main's import graph at
import-rfx time. Falsifier test now monkeypatches
rfx.geometry._pole_keying._pole_key (where _accumulate_pole_mask
resolves it). New regression lock: subprocess-pinned
'import rfx; from rfx.geometry import rasterize; callable(...)' names
the landmine so the next import refactor cannot silently re-trip it
(subprocess because in-process pytest import history can clobber the
attribute via the pre-existing runner imports; PYTHONPATH pins the
repo under test — the tutorial's own subprocess resolves the installed
rfx, which is why the red was CI-only).

Verified in-worktree: import rfx -> rasterize is <function>,
callable True (was <module>, False). Byte-identity unaffected: pure
code motion, helper bodies verbatim, concrete-pole key path untouched.

R3: memory=none-after-grep (import-graph fix; no physics path touched) | R2-attempts=0 | falsifier=subprocess callable lock in test file
